+++ src/AppBundle/Controller/OfferController.php
@@ -2,7 +2,7 @@
 /**
  * OfferController File Doc Comment
  *
- * PHP version 7.1
+ * PHP version 7.2
  *
  * @category OfferController
  * @package  Controller
@@ -10,10 +10,11 @@
  */
 namespace AppBundle\Controller;
 
+use AppBundle\Entity\Activity;
 use AppBundle\Entity\Offer;
+use AppBundle\Service\ManagementFeesService;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -31,79 +32,115 @@ class OfferController extends Controller
     /**
      * Lists all offer entities.
      *
-     * @Route("/",    name="offer_index")
-     * @Method("GET")
+     * @Route("/", methods={"GET"}, name="offer_index")
      *
      * @return Response A Response instance
      */
     public function indexAction()
     {
-        $em = $this->getDoctrine()->getManager();
-
-        $offers = $em->getRepository('AppBundle:Offer')->findAll();
-
-        return $this->render(
-            'offer/index.html.twig', array(
-            'offers' => $offers,
-            )
-        );
-    }
-
-    /**
-     * Creates a new offer entity.
-     *
-     * @param Request $request New posted info
-     *
-     * @Route("/new",  name="offer_new")
-     * @Method({"GET", "POST"})
-     *
-     * @return Response A Response instance
-     */
-    public function newAction(Request $request)
-    {
-        $offer = new Offer();
-        $form = $this->createForm('AppBundle\Form\OfferType', $offer);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
+        if ($this->getUser()->hasRole('ROLE_STRUCTURE')
+            && $this->getUser()->getOrganization()->getIsActive() === 1
+        ) {
             $em = $this->getDoctrine()->getManager();
-            $em->persist($offer);
-            $em->flush();
 
-            return $this->redirectToRoute(
-                'offer_show', array(
-                'id' => $offer->getId())
+            $event_activities = $em->getRepository('AppBundle:Activity')->findBy(
+                array(
+                    'organizationActivities' => $this->getUser()->getOrganization(),
+                    'type' => 'Évènement eSport'
+                )
+            );
+
+            $stream_activities = $em->getRepository('AppBundle:Activity')->findBy(
+                array(
+                    'organizationActivities' => $this->getUser()->getOrganization(),
+                    'type' => 'Activité de streaming'
+                )
+            );
+
+            $team_activities = $em->getRepository('AppBundle:Activity')->findBy(
+                array(
+                    'organizationActivities' => $this->getUser()->getOrganization(),
+                    'type' => 'Equipe eSport'
+                )
+            );
+
+            return $this->render(
+                'offer/index.html.twig', array(
+                    'event_activities' => $event_activities,
+                    'stream_activities' => $stream_activities,
+                    'team_activities' => $team_activities
+                )
             );
         }
-
-        return $this->render(
-            'offer/new.html.twig', array(
-            'offer' => $offer,
-            'form' => $form->createView(),
-            )
-        );
+        return $this->redirectToRoute('redirect');
     }
 
     /**
-     * Finds and displays a offer entity.
+     * Creates a new offer entity and get information for current activity.
      *
-     * @param Offer $offer The offer entity
+     * @param Request               $request     New offer posted with activity id
+     * @param Activity              $activity    New offer by activity
+     * @param ManagementFeesService $feesService Fees Calculation services
      *
-     * @Route("/{id}", name="offer_show")
-     * @Method("GET")
+     * @Route("/{id}/new", methods={"GET", "POST"}, name="offer_new")
      *
      * @return Response A Response instance
+     * @throws \Exception
      */
-    public function showAction(Offer $offer)
-    {
-        $deleteForm = $this->_createDeleteForm($offer);
+    public function newAction(Request $request, Activity $activity,
+        ManagementFeesService $feesService
+    ) {
+        if ($this->getUser()->hasRole('ROLE_STRUCTURE')
+            && $this->getUser()->getOrganization()->getIsActive() === 1
+        ) {
+            if ($this->getUser()->getOrganization()->getOrganizationActivity()->isEmpty()
+            ) {
+                $this->redirectToRoute('activity_new');
+            }
 
-        return $this->render(
-            'offer/show.html.twig', array(
-            'offer' => $offer,
-            'delete_form' => $deleteForm->createView(),
-            )
-        );
+            $offer = new Offer();
+            $form = $this->createForm('AppBundle\Form\OfferType', $offer);
+            $form->handleRequest($request);
+
+            $interval = new \DateInterval(
+                $this->getParameter('periode')
+            );
+            $fees = $feesService->getFees(
+                $offer->getAmount(), $offer->getFinalDeal()
+            );
+
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $offer = $offer
+                    ->setActivity($activity)
+                    ->setNameCanonical($offer->getName())
+                    ->setHandlingFee($fees)
+                    ->setDate($offer->getDate()->sub($interval))
+                    ->setPartnershipNb($form['partnershipNumber']->getData());
+                $em->persist($offer);
+                $em->flush();
+
+                $this->addFlash(
+                    'success',
+                    "Votre offre a bien été créée."
+                );
+
+                return $this->redirectToRoute(
+                    'offer_edit', array(
+                        'id' => $offer->getId()
+                    )
+                );
+            }
+
+            return $this->render(
+                'offer/new.html.twig', array(
+                    'offer' => $offer,
+                    'activity' => $activity,
+                    'form' => $form->createView()
+                )
+            );
+        }
+        return $this->redirectToRoute('redirect');
     }
 
     /**
@@ -112,62 +149,81 @@ class OfferController extends Controller
      * @param Request $request Edit posted info
      * @param Offer   $offer   The offer entity
      *
-     * @Route("/{id}/edit", name="offer_edit")
-     * @Method({"GET",      "POST"})
+     * @Route("/{id}/edit",
+     *     methods={"GET", "POST"}, name="offer_edit"
+     * )
      *
      * @return Response A Response instance
      */
     public function editAction(Request $request, Offer $offer)
     {
-        $deleteForm = $this->_createDeleteForm($offer);
-        $editForm = $this->createForm('AppBundle\Form\OfferType', $offer);
-        $editForm->handleRequest($request);
+        $user = $this->getUser();
+        if ($user->getOrganization()->getOrganizationActivity()->contains($offer->getActivity())
+        ) {
+            $partnershipNb = $offer->countPartnershipNumber();
+            $offer->setPartnershipNumber($partnershipNb);
 
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
+            $deleteForm = $this->_createDeleteForm($offer);
+            $editForm = $this->createForm('AppBundle\Form\OfferType', $offer);
+            $editForm->handleRequest($request);
 
-            return $this->redirectToRoute(
-                'offer_edit', array(
-                'id' => $offer->getId())
+            if ($editForm->isSubmitted() && $editForm->isValid()) {
+                $offer->setPartnershipNb($editForm['partnershipNumber']->getData());
+                $this->getDoctrine()->getManager()->persist($offer);
+                $this->getDoctrine()->getManager()->flush();
+
+                $this->addFlash(
+                    'success',
+                    "Vos modifications ont bien été prises en compte."
+                );
+
+                return $this->redirectToRoute(
+                    'dashboard_index', array(
+                        'id' => $offer->getId())
+                );
+            }
+
+            return $this->render(
+                'offer/edit.html.twig', array(
+                    'offer' => $offer,
+                    'edit_form' => $editForm->createView(),
+                    'delete_form' => $deleteForm->createView()
+                )
             );
         }
-
-        return $this->render(
-            'offer/edit.html.twig', array(
-            'offer' => $offer,
-            'edit_form' => $editForm->createView(),
-            'delete_form' => $deleteForm->createView(),
-            )
-        );
+        return $this->redirectToRoute('redirect');
     }
 
     /**
-     * Deletes a offer entity.
+     * Deletes an offer entity.
      *
      * @param Request $request Delete posted info
      * @param Offer   $offer   The offer entity
      *
-     * @Route("/{id}",   name="offer_delete")
-     * @Method("DELETE")
+     * @Route("/{id}", methods={"DELETE"}, name="offer_delete")
      *
      * @return Response A Response instance
      */
     public function deleteAction(Request $request, Offer $offer)
     {
-        $form = $this->_createDeleteForm($offer);
-        $form->handleRequest($request);
+        $user = $this->getUser();
+        if ($user->getOrganization()->getOrganizationActivity()->contains($offer->getActivity())
+        ) {
+            $form = $this->_createDeleteForm($offer);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            $em->remove($offer);
-            $em->flush();
+            if ($form->isSubmitted() && $form->isValid()) {
+                $em = $this->getDoctrine()->getManager();
+                $em->remove($offer);
+                $em->flush();
+            }
+            return $this->redirectToRoute('offer_index');
         }
-
-        return $this->redirectToRoute('offer_index');
+        return $this->redirectToRoute('redirect');
     }
 
     /**
-     * Creates a form to delete a offer entity.
+     * Creates a form to delete an offer entity.
      *
      * @param Offer $offer The offer entity
      *
@@ -178,8 +234,8 @@ class OfferController extends Controller
         return $this->createFormBuilder()
             ->setAction(
                 $this->generateUrl(
-                    'offer_delete',
-                    array('id' => $offer->getId())
+                    'offer_delete', array(
+                        'id' => $offer->getId())
                 )
             )
             ->setMethod('DELETE')
